+++ src/postoCombustivel/TesteBombaCombustivel.java
@@ -7,15 +7,15 @@ public class TesteBombaCombustivel {
     @Test
     void testeBombaConsultar() {
         Locale.setDefault(Locale.US);
-        BombaCombustivel bombaGasolina = new BombaCombustivel(10000f, 4.58f,
+        BombaCombustivel bombaGasolina = new BombaCombustivel(10000, 4.58,
                 "Gasolina");
         Assertions.assertEquals("Tipo de Combustivel: Gasolina\nValor por Litro: R$ 4.58\n" +
                 "Litros em Estoque: 10000.000\n", bombaGasolina.consultar());
     }
     @Test
     void testeBombaAlterartipoCombustivelAlcoolParaDiesel(){
-        BombaCombustivel bombaAlcoolParaDiesel = new BombaCombustivel(5000f,
-                4.00f, "Álcool" );
+        BombaCombustivel bombaAlcoolParaDiesel = new BombaCombustivel(5000,
+                4.00, "Álcool" );
 
         Assertions.assertEquals("Tipo de Combustivel alterado para: Diesel.",
                 bombaAlcoolParaDiesel.alterarTipoCombustivel('D'));
@@ -23,7 +23,7 @@ public class TesteBombaCombustivel {
     @Test
     void testeBombaAlteraLitrosEstoque(){
         Locale.setDefault(Locale.US);
-        BombaCombustivel bombaGasolina = new BombaCombustivel(5000f, 4.58f,
+        BombaCombustivel bombaGasolina = new BombaCombustivel(5000, 4.58,
                 "Gasolina");
 
         Assertions.assertEquals("Litros em Estoque alterado para: 4000.000.", bombaGasolina.alterarLitrosEstoque(4000f));
@@ -31,26 +31,26 @@ public class TesteBombaCombustivel {
     @Test
     void testeBombaGasolinaAlteraValorPorlitro4e58Para4e40(){
         Locale.setDefault(Locale.US);
-        BombaCombustivel bombaGasolina = new BombaCombustivel(5000f, 4.58f,
+        BombaCombustivel bombaGasolina = new BombaCombustivel(5000, 4.58,
                 "Gasolina");
-        float valorPorlitro = 4.40f;
+        double valorPorlitro = 4.40;
         Assertions.assertEquals("Valor por Litro alterado para: R$ 4.40.", bombaGasolina.alterarValorPorLitro(valorPorlitro));
     }
     @Test
     void testeBombaGasolinaAbastecerPorLitro(){
         Locale.setDefault(Locale.US);
-        BombaCombustivel bombaGasolina = new BombaCombustivel(5000f, 4.40f,
+        BombaCombustivel bombaGasolina = new BombaCombustivel(5000, 4.40,
                 "Gasolina");
 
         Assertions.assertEquals("8.80", bombaGasolina.abastecerPorLitro(2));
-        Assertions.assertEquals(4998f, bombaGasolina.quantidadeDeCombustivel);
+        Assertions.assertEquals(4998, bombaGasolina.quantidadeDeCombustivel);
     }
     @Test
     void testeBombaGasolinaAbastecerPorValor(){
         Locale.setDefault(Locale.US);
-        BombaCombustivel bombaGasolina = new BombaCombustivel(5000f, 4.40f,
+        BombaCombustivel bombaGasolina = new BombaCombustivel(5000, 4.40,
                 "Gasolina");
-        Assertions.assertEquals("8.80", bombaGasolina.abastecerPorValor(8.80f));
-        Assertions.assertEquals(4998f, bombaGasolina.quantidadeDeCombustivel);
+        Assertions.assertEquals("8.80", bombaGasolina.abastecerPorValor(8.80));
+        Assertions.assertEquals(4998, bombaGasolina.quantidadeDeCombustivel);
     }
 }
